requirements.txt
flask
mysql-connector-python
qrcode[pil]
pillow
werkzeug